+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [#3455](https://github.com/bbatsov/rubocop/issues/3455): Add new `Rails/DynamicFindBy` cop. ([@pocke][])
 * [#3542](https://github.com/bbatsov/rubocop/issues/3542): Add a configuration option, `IgnoreCopDirectives`, to `Metrics/LineLength` to stop cop directives (`# rubocop:disable Metrics/AbcSize`) from being counted when considering line length. ([@jmks][])
 * Add new `Rails/DelegateAllowBlank` cop. ([@connorjacobsen][])
+* Add new `Style/MultilineMemoization` cop. ([@drenmi][])
 
 ### Bug fixes